Move RealizedAssembly into code.ts since that's the only place it's used. Various import formatting fixes as well.
